Use ptrs to reduce number of copies
Using pprof, it was clear that unionFailOutput is making a lot of copies
which causes high memory usage and hurts cpu runtime. I switched it to
use ptrs even though I had previously expected no copying to be done
since we were dealing with slices. I need to think about a better way to
do this, perhaps I should use indices and reference the index of the
string in the original pattern slice. For this to word, the original
slice order must be unmodified, this might be problematic due to the
sorting I do, futher investigation is needed.